+++ R/overhaul.R
@@ -1,3 +1,5 @@
+utils::globalVariables(c("start"))
+
 #' Edges
 #' 
 #' Get edges from data.frame of tweets.
@@ -6,7 +8,18 @@
 #' @param tweets Column containing tweets.
 #' @param source Author of tweets.
 #' @param id tweets unique id.
+#' @param hashtags Column containing hashtags.
+#' @param tl Set to \code{TRUE} to convert hashtags to lower case.
 #' @param ... any other column name, see examples.
+#' 
+#' @section Functions:
+#' \itemize{
+#'   \item{\code{gt_edges} - Build networks of users.}
+#'   \item{\code{gt_edges_hash} - Build networks of users to hashtags.}
+#' }
+#' 
+#' @details The \code{tl} arguments stands for \code{\link{tolower}} and allows converting the #hashtags to lower case as 
+#' these often duplicated, i.e.: #python #Python.
 #' 
 #' @examples 
 #' # simulate dataset
@@ -16,6 +29,7 @@
 #'   screen_name = c("me", "him"),
 #'   retweet_count = c(19, 5),
 #'   status_id = c(1, 2),
+#'   hashtags = c("rstats", "Python"),
 #'   stringsAsFactors = FALSE
 #' )
 #'
@@ -23,15 +37,20 @@
 #'   gt_edges(text, screen_name, status_id)
 #'   
 #' tweets %>% 
-#'   gt_edges_(RT = "retweet_count") # metadata
-#'   
+#'   gt_edges_(RT = "retweet_count") %>% 
+#'   gt_nodes()
+#'  
+#' tweets %>% 
+#'   gt_edges_hash(hashtags, screen_name) %>% 
+#'   gt_nodes()
+#'     
 #' @return An object of class \code{graphTweets}.
 #' 
 #' @rdname edges
 #' @export
 gt_edges <- function(data, tweets, source, id, ...){
-  if(missing(data))
-    stop("missing data", call. = FALSE)
+  if(missing(data) || missing(tweets) || missing(source) || missing(id))
+    stop("missing data, tweets, source, or id", call. = FALSE)
   tweets <- dplyr::enquo(tweets)
   source <- dplyr::enquo(source)
   id <- deparse(substitute(id))
@@ -40,7 +59,7 @@ gt_edges <- function(data, tweets, source, id, ...){
 
 #' @rdname edges
 #' @export
-gt_edges_ <- function(data, tweets = "text", source = "screen_name",  id = "status_id", ...){
+gt_edges_ <- function(data, tweets = "text", source = "screen_name", id = "status_id", ...){
   
   if(missing(data))
     stop("missing data", call. = FALSE)
@@ -79,10 +98,49 @@ gt_edges_ <- function(data, tweets = "text", source = "screen_name",  id = "stat
       target = "handles",
       ...
       ) %>% 
-    dplyr::as_tibble()-> edges
+    dplyr::as_tibble() %>% 
+    dplyr::filter(
+      source != "",
+      target != ""
+    ) -> edges
   
   construct(data, edges, NULL)
   
+}
+
+#' @rdname edges
+#' @export
+gt_edges_hash <- function(data, hashtags, source, ..., tl = TRUE){
+  if(missing(data) || missing(hashtags) || missing(source))
+    stop("missing data, hashtags, or source", call. = FALSE)
+  hashtags <- deparse(substitute(hashtags))
+  source <- deparse(substitute(source))
+  gt_edges_hash_(data, hashtags, source, ..., tl = tl)
+}
+
+#' @rdname edges
+#' @export
+gt_edges_hash_ <- function(data, hashtags = "hashtags", source = "screen_name", ..., tl = TRUE){
+  
+  if(missing(data))
+    stop("missing data", call. = FALSE)
+  
+  edges <- data %>% 
+    dplyr::select_(hashtags, source, ...) %>% 
+    tidyr::unnest_("hashtags") %>% 
+    dplyr::mutate(
+      hashtags = dplyr::case_when(
+        tl == TRUE ~ tolower(hashtags),
+        TRUE ~ hashtags
+      )
+    ) %>% 
+    dplyr::group_by_("screen_name", "hashtags", ...) %>% 
+    dplyr::count() %>% 
+    dplyr::ungroup() %>% 
+    dplyr::mutate(hashtags = paste0("#", hashtags))
+  
+  names(edges)[1:3] <- c("source", "target", "n_tweets")
+  construct(data, edges, NULL)
 }
 
 #' Nodes
@@ -107,7 +165,7 @@ gt_edges_ <- function(data, tweets = "text", source = "screen_name",  id = "stat
 #'   gt_edges(text, screen_name, status_id) %>% 
 #'   gt_nodes() -> net
 #'   
-#' @return An object of class \code{graphTweets}.
+#' @return An object of class \code{graphTweets}, adds \code{nodes}.
 #' 
 #' @export
 gt_nodes <- function(gt, meta = FALSE){
@@ -117,10 +175,25 @@ gt_nodes <- function(gt, meta = FALSE){
   
   test_input(gt)
   
-  nodes <- unique(c(gt[["edges"]][["source"]], gt[["edges"]][["target"]]))
+  nodes <- c(gt[["edges"]][["source"]], gt[["edges"]][["target"]])
+  
+  if("n_tweets" %in% names(gt[["edges"]]))
+    type <- c(
+      rep("user", nrow(gt[["edges"]])),
+      rep("hashtag", nrow(gt[["edges"]]))
+    )
+  else
+    type <- "user"
+  
   nodes <- dplyr::tibble(
-    nodes = nodes
-  )
+    nodes = nodes,
+    type = type
+  ) %>% 
+    dplyr::group_by(nodes, type) %>% 
+    dplyr::summarise(
+      n_edges = n()
+    ) %>% 
+    dplyr::ungroup()
   
   if(isTRUE(meta)){
     usr <- rtweet::users_data(gt$tweets)
@@ -218,8 +291,9 @@ gt_graph <- function(gt){
 #'
 #' tweets %>% 
 #'   gt_edges(text, screen_name, status_id, "created_at") %>% 
-#'   gt_nodes(TRUE) %>% 
-#'   gt_dyn() -> net
+#'   gt_nodes() %>% 
+#'   gt_dyn() %>% 
+#'   gt_collect() -> net
 #' }
 #' 
 #' @rdname dyn
@@ -247,10 +321,12 @@ gt_dyn <- function(gt, lifetime = Inf){
     dplyr::group_by(source) %>% 
     unique() %>% 
     dplyr::summarise(
-      start = min(created_at),
-      end = max(created_at)
+      start = min(created_at)
     ) %>% 
     dplyr::ungroup() %>% 
+    dplyr::mutate(
+      end = max(start)
+    ) %>% 
     dplyr::inner_join(gt[["nodes"]], by = c("source" = "nodes"))
   
   if(nrow(nodes) != nrow(gt[["nodes"]]))