+++ NEWS.md
@@ -1,3 +1,8 @@
+# graphTweets 0.4.1
+
+* `gt_nodes` returns number of `n_edges`, the number of edges the node is present in.
+* `gt_edges_hash` and respective escape hatch added! 
+
 # graphTweets 0.4.0
 
 Major release: overhaul to 1) make computations much faster, 2) make the whole package more tidyverse friendly and 3) switch to `rtweet` as main source.